+++ CHANGELOG.md
@@ -9,8 +9,9 @@ Repository: **<git+https://github.com/vladmandic/human.git>**
 
 ## Changelog
 
-### **HEAD -> main** 2021/03/16 mandic00@live.com
+### **HEAD -> main** 2021/03/17 mandic00@live.com
 
+- cleanup
 
 ### **1.1.7** 2021/03/16 mandic00@live.com